requirements.txt
# Dependencias principales
requests>=2.28.0
PyGithub>=1.58.0
python-dotenv>=0.19.0
PyYAML>=6.0

# Análisis de texto y NLP
nltk>=3.8
textblob>=0.17.1
vaderSentiment>=3.3.2
spacy>=3.4.0
transformers>=4.20.0
torch>=1.12.0

# Análisis de datos
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.1.0

# Visualización
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.10.0

# Web framework
flask>=2.2.0
flask-cors>=3.0.10

# Base de datos
sqlalchemy>=1.4.0
alembic>=1.8.0

# Utilidades
click>=8.1.0
tqdm>=4.64.0
python-dateutil>=2.8.0

# Testing (desarrollo)
pytest>=7.1.0
pytest-cov>=3.0.0
black>=22.6.0
flake8>=5.0.0